Move brick to bin/glusterd2/brick
