+++ bin/glusterd2/commands/volumes/brickutils.go
@@ -5,7 +5,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/gluster/glusterd2/brick"
+	"github.com/gluster/glusterd2/bin/glusterd2/brick"
 	"github.com/gluster/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/utils"
