+++ bin/glusterd2/commands/volumes/volume-expand.go
@@ -3,12 +3,12 @@ package volumecommands
 import (
 	"net/http"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/brick"
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
 	"github.com/gluster/glusterd2/bin/glusterd2/volume"
-	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/pkg/api"
 	"github.com/gluster/glusterd2/pkg/errors"
 	"github.com/gluster/glusterd2/pkg/utils"
