+++ bin/glusterd2/volgen/volgen.go
@@ -8,9 +8,9 @@ import (
 	"path"
 	"strings"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/brick"
 	"github.com/gluster/glusterd2/bin/glusterd2/store"
 	"github.com/gluster/glusterd2/bin/glusterd2/volume"
-	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/pkg/utils"
 )
