+++ bin/glusterd2/volume/struct.go
@@ -9,9 +9,9 @@ import (
 	"net/http"
 	"path/filepath"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/brick"
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/bin/glusterd2/peer"
-	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/pkg/utils"
 
 	"github.com/pborman/uuid"
